Use anonymous mapping instead of dropVM
This is a suggested alternative implementation, in the form of a PR.

I'm just using this as a way of having a discussion with code attached.

I noticed that you drop vm caches. This is undesirable, since it means throwing
away all of the page caches of the entire system. I've used systems where I had
hundreds of gigabytes of useful data in those caches, so I really would prefer
to keep them.

But further, I didn't understand from the commit log why it was necessary in
order to get a page fault. We can just as well get a page fault by mapping an
anonymous page, no? I've done that in this PR and it seems to pass. If there is
a good reason to map a named file and drop caches, I'd love to understand it.