Drop use of NOSPLIT on doEnableRunDisable
This fixes the SampleTracepointStack test, because the kernel stack trace
wasn't able to 'see' the caller of doEnableRunDisable since we weren't
maintaining the frame pointer.

My original rationale for using NOSPLIT was to reduce the overhead of calling
doEnableRunDisable, but that doesn't make sense since we're called with the
counters inhibited anyway, so any stack splitting code won't be counted.

The thing which needs to be NOSPLIT is the `f` to be called, which is under the
control of the user.